requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
markdownify==0.11.6
urllib3==2.0.7
html2text==2020.1.16
firecrawl-py==0.0.16
python-dotenv==1.0.0
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.24.0 